backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy[asyncio]>=2.0
asyncpg
psycopg2-binary
alembic
greenlet
supabase>=2.0.0,<2.10.0
pandas
requests
httpx
beautifulsoup4
apscheduler
geopandas
scipy
scikit-learn
xgboost
joblib
google-generativeai
